Move tile classes to tile module
