+++ android/app/src/main/AndroidManifest.xml
@@ -73,7 +73,7 @@
             Tile services must be exported and protected by the bind tile permission
             (android.permission.BIND_QUICK_SETTINGS_TILE).
         -->
-        <service android:name="net.mullvad.mullvadvpn.service.MullvadTileService"
+        <service android:name="net.mullvad.mullvadvpn.tile.MullvadTileService"
                  android:exported="true"
                  android:permission="android.permission.BIND_QUICK_SETTINGS_TILE"
                  android:label="@string/toggle_vpn"
