+++ android/tile/src/main/kotlin/net/mullvad/mullvadvpn/tile/TileConstants.kt
@@ -1,0 +1,6 @@
+package net.mullvad.mullvadvpn.tile
+
+internal const val VPN_SERVICE_PACKAGE = "net.mullvad.mullvadvpn"
+internal const val VPN_SERVICE_CLASS = "$VPN_SERVICE_PACKAGE.service.MullvadVpnService"
+internal const val KEY_CONNECT_ACTION = "$VPN_SERVICE_PACKAGE.connect_action"
+internal const val KEY_DISCONNECT_ACTION = "$VPN_SERVICE_PACKAGE.disconnect_action"
